TASK: [gw] Move client modules to control namespace
