+++ crates/gateway/src/control/client/controller/signaling.rs
@@ -15,11 +15,11 @@ use trust0_common::logging::error;
 use trust0_common::net::tls_server::conn_std;
 use trust0_common::{sync, target};
 
-use crate::client::controller::signaling::certificate_reissue::CertReissuanceProcessor;
-use crate::client::controller::signaling::proxy_connections::ProxyConnectionsProcessor;
-use crate::client::controller::{ChannelProcessor, ControlPlane};
-use crate::client::device::Device;
 use crate::config::AppConfig;
+use crate::control::client::controller::signaling::certificate_reissue::CertReissuanceProcessor;
+use crate::control::client::controller::signaling::proxy_connections::ProxyConnectionsProcessor;
+use crate::control::client::controller::{ChannelProcessor, ControlPlane};
+use crate::control::client::device::Device;
 use crate::service::manager::ServiceMgr;
 
 pub const EVENT_LOOP_CYCLE_DELAY_MSECS: u64 = 6_000;
@@ -215,8 +215,8 @@ pub trait SignalingEventHandler {
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use crate::client::controller::tests::{create_device, MockChannelProc};
     use crate::config;
+    use crate::control::client::controller::tests::{create_device, MockChannelProc};
     use crate::repository::access_repo::tests::MockAccessRepo;
     use crate::repository::role_repo::tests::MockRoleRepo;
     use crate::repository::service_repo::tests::MockServiceRepo;
