+++ crates/gateway/src/control/client/device.rs
