+++ crates/gateway/src/control/mod.rs
@@ -1,0 +1,1 @@
+pub mod client;
